chore: 剥离中文数字转换逻辑
